Drafted base-compositor and vive-compositor cabal executables; eliminated simple-compositor.
Commentary:

1. **Naming Hierarchy.** The `BaseCompositor` vs. `ViveCompositor` distinction
might not be the right way to divide things conceptually (@krakrjak); for now it keeps
things moving though.
2. **Cabal flags are still needed.** Right now, the base-compositor still depends on libOsvrClientKit.
3. **DBus code removed.** I assumed this was no longer needed (perhaps stupidly
since we don't have a running vive-compositor yet; apologies if this was a
stupid cleaning move).
4. **Weston.hs is still huge.** I tried to transform the old
`Compositor.Weston` into `BaseCompositor.Weston` and
`BaseCompositor.BaseCompositor`; however, I entered a cyclical dependency
nightmare and so just kept it all together to move on with it.